services/fastapi/requirements.txt
fastapi>=0.104,<1.0
uvicorn[standard]>=0.24,<1.0
pydantic>=2.4,<3.0
pydantic-settings>=2.0,<3.0
httpx>=0.25,<1.0
redis>=5.0,<6.0
python-jose[cryptography]>=3.3,<4.0
python-multipart>=0.0.6,<1.0
celery[redis]>=5.3,<6.0
